Passing -Weverything -Wno-qual-cast flags now

+++ src/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rhallste <rhallste@student.42.us.org>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/09/19 11:54:10 by rhallste          #+#    #+#             */
-/*   Updated: 2017/11/20 18:21:19 by rhallste         ###   ########.fr       */
+/*   Updated: 2017/11/22 10:11:55 by rhallste         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (s[i])
